logout modify
로그아웃 반복이 나와서 라우터 방어 코드 수정하면서 로그아웃 수정하고 관련 코드 수정

로그아웃은 단순히 서비스 유지 나머지 코드에서 초기화 로직 수정함(나중에 리팩터링 해야함)

+++ frontend-master/src/utils/RequireAuth.jsx
@@ -3,6 +3,8 @@ import React from "react";
 import { Navigate, Outlet } from "react-router-dom";
 import { getAccessToken, removeAccessToken } from "./TokenManager";
 import { logout } from "@/services/LogoutService";
+import { useEffect } from "react";
+import { useRef } from "react";
 
 /**
  * 로그인 사용자 정보가 없으면 로그인 페이지로 이동하는 ProtectedRoute
@@ -11,15 +13,26 @@ import { logout } from "@/services/LogoutService";
 export default function RequireAuth() {
   const { userInfo, setDefaultUser } = useLoginUserInfoStore();
   const token = getAccessToken();
+  const hasLoggedOut = useRef(false); // 로그아웃 중복 방지용
 
   const isAuth = token && userInfo?.userId;
 
+  useEffect(() => {
+    if (!isAuth && !hasLoggedOut.current) {
+      hasLoggedOut.current = true;
+      (async () => {
+        await logout();
+        removeAccessToken();
+        sessionStorage.clear();
+        localStorage.clear();
+        setDefaultUser();
+      })();
+    }
+  }, [isAuth, setDefaultUser]);
+
   if (!isAuth) {
-    logout();
-    setDefaultUser();
-    sessionStorage.clear();
-    localStorage.clear();
     return <Navigate to="/user/login" replace />;
   }
+
   return <Outlet />;
 }
